backend/requirements.txt
duckduckgo-search>=6.0.0
gspread>=6.0.0
oauth2client>=4.1.3
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
sse-starlette>=1.6.1
openai>=1.0.0
